Move script-related endpoint code to scripts file
