+++ ee/server/service/hosts.go
@@ -2,11 +2,7 @@ package service
 
 import (
 	"context"
-	"net/http"
-	"time"
 
-	"github.com/fleetdm/fleet/v4/server/authz"
-	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 )
 
@@ -22,145 +18,4 @@ func (svc *Service) HostByIdentifier(ctx context.Context, identifier string, opt
 	opts.IncludeCVEScores = true
 	opts.IncludePolicies = true
 	return svc.Service.HostByIdentifier(ctx, identifier, opts)
-}
-
-func (svc *Service) RunHostScript(ctx context.Context, request *fleet.HostScriptRequestPayload, waitForResult time.Duration) (*fleet.HostScriptResult, error) {
-	const maxPendingScriptAge = time.Minute // any script older than this is not considered pending anymore on that host
-
-	// must load the host to get the team (cannot use lite, the last seen time is
-	// required to check if it is online) to authorize with the proper team id.
-	// We cannot first authorize if the user can list hosts, in case we
-	// eventually allow a write-only role (e.g. gitops).
-	host, err := svc.ds.Host(ctx, request.HostID)
-	if err != nil {
-		// if error is because the host does not exist, check first if the user
-		// had access to run a script (to prevent leaking valid host ids).
-		if fleet.IsNotFound(err) {
-			if err := svc.authz.Authorize(ctx, &fleet.HostScriptResult{}, fleet.ActionWrite); err != nil {
-				return nil, err
-			}
-		}
-		svc.authz.SkipAuthorization(ctx)
-		return nil, ctxerr.Wrap(ctx, err, "get host lite")
-	}
-	if err := svc.authz.Authorize(ctx, &fleet.HostScriptResult{TeamID: host.TeamID}, fleet.ActionWrite); err != nil {
-		return nil, err
-	}
-
-	if err := fleet.ValidateHostScriptContents(request.ScriptContents); err != nil {
-		return nil, fleet.NewInvalidArgumentError("script_contents", err.Error())
-	}
-
-	// host must be online
-	if host.Status(time.Now()) != fleet.StatusOnline {
-		return nil, fleet.NewInvalidArgumentError("host_id", fleet.RunScriptHostOfflineErrMsg)
-	}
-
-	pending, err := svc.ds.ListPendingHostScriptExecutions(ctx, request.HostID, maxPendingScriptAge)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "list host pending script executions")
-	}
-	if len(pending) > 0 {
-		return nil, fleet.NewInvalidArgumentError(
-			"script_contents", fleet.RunScriptAlreadyRunningErrMsg,
-		).WithStatus(http.StatusConflict)
-	}
-
-	// create the script execution request, the host will be notified of the
-	// script execution request via the orbit config's Notifications mechanism.
-	script, err := svc.ds.NewHostScriptExecutionRequest(ctx, request)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "create script execution request")
-	}
-	script.Hostname = host.DisplayName()
-
-	asyncExecution := waitForResult <= 0
-
-	err = svc.ds.NewActivity(
-		ctx,
-		authz.UserFromContext(ctx),
-		fleet.ActivityTypeRanScript{
-			HostID:            host.ID,
-			HostDisplayName:   host.DisplayName(),
-			ScriptExecutionID: script.ExecutionID,
-			Async:             asyncExecution,
-		},
-	)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "create activity for script execution request")
-	}
-
-	if asyncExecution {
-		// async execution, return
-		return script, nil
-	}
-
-	ctx, cancel := context.WithTimeout(ctx, waitForResult)
-	defer cancel()
-
-	// if waiting for a result times out, we still want to return the script's
-	// execution request information along with the error, so that the caller can
-	// use the execution id for later checks.
-	timeoutResult := script
-	checkInterval := time.Second
-	after := time.NewTimer(checkInterval)
-	for {
-		select {
-		case <-ctx.Done():
-			return timeoutResult, ctx.Err()
-		case <-after.C:
-			result, err := svc.ds.GetHostScriptExecutionResult(ctx, script.ExecutionID)
-			if err != nil {
-				// is that due to the context being canceled during the DB access?
-				if ctxErr := ctx.Err(); ctxErr != nil {
-					return timeoutResult, ctxErr
-				}
-				return nil, ctxerr.Wrap(ctx, err, "get script execution result")
-			}
-			if result.ExitCode != nil {
-				// a result was received from the host, return
-				result.Hostname = host.DisplayName()
-				return result, nil
-			}
-
-			// at a second to every attempt, until it reaches 5s (then check every 5s)
-			if checkInterval < 5*time.Second {
-				checkInterval += time.Second
-			}
-			after.Reset(checkInterval)
-		}
-	}
-}
-
-func (svc *Service) GetScriptResult(ctx context.Context, execID string) (*fleet.HostScriptResult, error) {
-	scriptResult, err := svc.ds.GetHostScriptExecutionResult(ctx, execID)
-	if err != nil {
-		if fleet.IsNotFound(err) {
-			if err := svc.authz.Authorize(ctx, &fleet.HostScriptResult{}, fleet.ActionRead); err != nil {
-				return nil, err
-			}
-		}
-		svc.authz.SkipAuthorization(ctx)
-		return nil, ctxerr.Wrap(ctx, err, "get script result")
-	}
-
-	host, err := svc.ds.HostLite(ctx, scriptResult.HostID)
-	if err != nil {
-		// if error is because the host does not exist, check first if the user
-		// had access to run a script (to prevent leaking valid host ids).
-		if fleet.IsNotFound(err) {
-			if err := svc.authz.Authorize(ctx, &fleet.HostScriptResult{}, fleet.ActionRead); err != nil {
-				return nil, err
-			}
-		}
-		svc.authz.SkipAuthorization(ctx)
-		return nil, ctxerr.Wrap(ctx, err, "get host lite")
-	}
-	if err := svc.authz.Authorize(ctx, &fleet.HostScriptResult{TeamID: host.TeamID}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
-
-	scriptResult.Hostname = host.DisplayName()
-
-	return scriptResult, nil
 }
